+++ src/routes/graphql/types/memberType-type.ts
@@ -1,0 +1,18 @@
+import { GraphQLObjectType, GraphQLFloat, GraphQLInt, GraphQLEnumType } from 'graphql';
+
+export const MemberTypeType = new GraphQLObjectType({
+  name: 'MemberType',
+  fields: () => ({
+    id: { type: MemberTypeId },
+    discount: { type: GraphQLFloat },
+    postsLimitPerMonth: { type: GraphQLInt },
+  }),
+});
+
+export const MemberTypeId = new GraphQLEnumType({
+  name: 'MemberTypeId',
+  values: {
+    basic: { value: 'basic' },
+    business: { value: 'business' },
+  },
+});
