+++ src/routes/graphql/types/post-type.ts
@@ -1,0 +1,11 @@
+import { GraphQLObjectType, GraphQLString } from 'graphql';
+import { UUIDType } from './uuid.js';
+
+export const PostType = new GraphQLObjectType({
+  name: 'Post',
+  fields: () => ({
+    id: { type: UUIDType },
+    title: { type: GraphQLString },
+    content: { type: GraphQLString },
+  }),
+});
